requirements.txt
asgiref==3.8.1
attrs==24.2.0
az-iranian-bank-gateways==2.0.8
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
Django==5.0.7
django-comments-dab==2.8.0
django-taggit==6.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
isodate==0.6.1
lxml==5.3.0
MarkupSafe==2.1.5
moviepy==1.0.3
numpy==2.0.1
pillow==10.4.0
platformdirs==4.2.2
proglog==0.1.10
pycryptodome==3.20.0
pytz==2024.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.2
zeep==4.2.1
